backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bleach==6.1.0
blinker==1.4
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.2
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cytoolz==0.12.3
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
docopt==0.6.2
eth-account==0.11.2
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-tester==0.11.0b2
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
frozenlist==1.4.1
hexbytes==0.3.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
install==1.3.5
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lru-dict==1.2.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.10.0
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.0
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreq==0.4
pipreqs==0.5.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-ecc==7.0.1
py-evm==0.10.1b1
py-geth==4.4.0
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
pyunormalize==15.1.0
PyYAML==5.4.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.1
rlp==4.0.1
rpds-py==0.18.1
safe-pysha3==1.0.4
SecretStorage==3.3.1
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
ssh-import-id==5.11
stack-data==0.6.3
systemd-python==234
tinycss2==1.3.0
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
trie==3.0.1
typing_extensions==4.11.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
virtualenv==20.26.2
wadllib==1.3.6
wcwidth==0.2.13
web3==6.19.0
webencodings==0.5.1
websockets==12.0
yarg==0.1.9
yarl==1.9.4
zipp==1.0.0
